+++ src/main/java/ddd/caffeine/ratrip/module/feign/auth/kakao/KakaoUserApiClient.java
@@ -1,10 +1,10 @@
-package ddd.caffeine.ratrip.module.external;
+package ddd.caffeine.ratrip.module.feign.auth.kakao;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 
-import ddd.caffeine.ratrip.module.external.dto.KakaoProfileResponse;
+import ddd.caffeine.ratrip.module.feign.auth.kakao.dto.KakaoProfileResponse;
 
 @FeignClient(name = "kakaoApiClient", url = "https://kapi.kakao.com")
 public interface KakaoUserApiClient {
