+++ src/main/java/ddd/caffeine/ratrip/module/feign/auth/kakao/dto/KakaoProfileResponse.java
@@ -1,4 +1,4 @@
-package ddd.caffeine.ratrip.module.external.dto;
+package ddd.caffeine.ratrip.module.feign.auth.kakao.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
