+++ src/main/java/ddd/caffeine/ratrip/module/feign/place/domain/kakao/KakaoFeignClient.java
@@ -1,4 +1,4 @@
-package ddd.caffeine.ratrip.module.feign.domain.place.kakao;
+package ddd.caffeine.ratrip.module.feign.place.domain.kakao;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
